+++ src/elexclarity/cli.py
@@ -5,13 +5,24 @@ import click
 from elexclarity.client import ElectionsClient
 from elexclarity.convert import convert
 
+
+class StringListParamType(click.ParamType):
+    name = "stringlist"
+
+    def convert(self, value, param, ctx):
+        return value.split(",")
+
+
 BASE_URL = os.environ.get('CLARITY_API_BASE_URL', 'https://results.enr.clarityelections.com/')
+STRING_LIST = StringListParamType()
 
 
 @click.command()
 @click.argument('electionID', type=click.INT)
 @click.argument('statePostal', type=click.STRING)
 @click.option('--filename', type=click.Path(exists=True), help='Specify data file instead of making HTTP request')
+@click.option('--countyMapping', 'countyMapping', default={}, help='Specify county mapping')
+@click.option('--officeID', 'officeID', help='The office ID(s) to process', type=STRING_LIST)
 @click.option('--level', help='Specify the subunit type/reporting level for results', default='county', type=click.Choice([
     'county',
     'precinct',
@@ -22,15 +33,15 @@ BASE_URL = os.environ.get('CLARITY_API_BASE_URL', 'https://results.enr.clarityel
     'settings',
     'results'
 ]))
-@click.option('--resultsBy', 'resultsBy', default='candidate', type=click.Choice([
-    'candidate',
-    'party',
+@click.option('--voteCompletionMode', 'voteCompletionMode', default='percentReporting', type=click.Choice([
+    'percentReporting',
+    'voteTypes'
 ]))
 @click.option('--style', default='default', type=click.Choice([
     'default',
     'raw'
 ]))
-def cli(electionid, statepostal, filename=None, style="default", outputType="results", resultsBy='candidate', **kwargs):
+def cli(electionid, statepostal, filename=None, countyMapping={}, outputType="results", **kwargs):
     """
     This tool accepts an election ID (e.g. 105369) and a state postal code (e.g. GA)
     and the options below and outputs formatted elections data. If a filename is provided,
@@ -41,16 +52,16 @@ def cli(electionid, statepostal, filename=None, style="default", outputType="res
     > elexclarity 105369 GA --outputType=summary
     > elexclarity 105369 GA --outputType=settings
     > elexclarity 105369 GA --level=precinct
-    > elexclarity 105369 GA --level=state
+    > elexclarity 105369 GA --level=county
     """
     if filename:
         # load races from local file
-        with click.open_file(filename) as f:
+        with click.open_file(filename) as results_file:
+            # load these into a list as that's what convert expects
             if ".json" in filename:
-                result = json.load(f)
+                result = [json.load(results_file)]
             if ".xml" in filename:
-                with open(filename) as f:
-                    result = f.read()
+                result = [results_file.read()]
     else:
         api_client = ElectionsClient(BASE_URL)
         if outputType == "summary":
@@ -60,8 +71,8 @@ def cli(electionid, statepostal, filename=None, style="default", outputType="res
         else:
             result = api_client.get_results(electionid, statepostal, **kwargs)
 
-    if style == "raw":
-        return result
+    if isinstance(countyMapping, str):
+        countyMapping = json.loads(countyMapping)
 
-    result = convert(result, statepostal=statepostal, outputType=outputType, **kwargs)
+    result = convert(result, statepostal, outputType=outputType, countyMapping=countyMapping, **kwargs)
     print(json.dumps(result, indent=2))