+++ src/elexclarity/convert.py
@@ -1,34 +1,59 @@
-from collections import defaultdict
-
-import xmltodict
-from elexstatic import STATE_COUNTIES
-from slugify import slugify
-
-from elexclarity.formatters import ClarityXMLConverter
+from elexclarity.formatters.results import ClarityDetailXMLConverter
+from elexclarity.formatters.settings import ClaritySettingsConverter
 
 
-def convert(data, statepostal=None, level=None, outputType="results", style="default", resultsBy=None, **kwargs):
+def convert(
+    data,
+    statepostal,
+    outputType="results",
+    style="default",
+    countyMapping=None,
+    officeID=None,
+    voteCompletionMode=None,
+    **kwargs
+):
     """
     The entry point for formatting Clarity results data.
     """
-    # TODO: Data formatting/conversion logic for settings and summary
-    if outputType == "summary" or outputType == "settings":
-        # Returns raw data for now
+
+    office_id = officeID
+    vote_completion_mode = voteCompletionMode
+
+    if office_id and isinstance(office_id, str):
+        office_id = office_id.split(",")
+
+    if style == "raw" or outputType == "summary":
         return data
 
-    if type(data) == list:
-        data = [xmltodict.parse(i, attr_prefix="")["ElectionResult"] for i in data]
-    else:
-        data = [xmltodict.parse(data, attr_prefix="")["ElectionResult"]]
+    if outputType == "settings":
+        return ClaritySettingsConverter(statepostal, county_lookup=countyMapping).convert(
+            data,
+            office_id=office_id,
+            **kwargs
+        )
 
-    if level == "precinct" or level == "county":
-        county_fips_lookup = {v["name"]: k for k, v in STATE_COUNTIES[statepostal].items()}
-        converter = ClarityXMLConverter(county_lookup=county_fips_lookup)
-        results = [converter.transform_result_object(i, level=level) for i in data]
+    if outputType == "results":
+        converter = ClarityDetailXMLConverter(
+            statepostal,
+            county_lookup=countyMapping
+        )
 
-        if len(results) > 1:
+        if isinstance(data, list):
+            results = {}
+            for sub_result in data:
+                results.update(converter.convert(
+                    sub_result,
+                    vote_completion_mode=vote_completion_mode,
+                    office_id=office_id,
+                    **kwargs
+                ))
             return results
+        else:
+            return converter.convert(
+                data,
+                vote_completion_mode=vote_completion_mode,
+                office_id=office_id,
+                **kwargs
+            )
 
-        return results[0]
-
-    raise Exception(f"The {level} Clarity formatter is not implemented yet")
+    raise Exception(f"The {outputType} Clarity formatter is not implemented yet")