+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.0.2] - 2021-01-04
+
+- [#6] Adds a timestamp to the race settings JSON output
+- [#7] Adds an option for a map to convert race names
+- [#8] Changes to election settings handling
+- [#9] Allow filtering by officeID in formatting step
+- [#10] Include county name in the race settings JSON output
+- [#11] README updates and CLI cleanup
+- [#12] Adds a new `voteCompletionMode` CLI option
+- [#13] Updates to use `race_type=G` for runoffs
+
+## [0.0.1-beta.2]
+
+- Adds a `voteCompletionMode` parameter to switch between using the `percentReporting` field vs vote counts by type to determine whether a precinct is fully reporting
+
+## [0.0.1-beta.1]
+
+- Initial beta release with county-level settings formatting and precinct, county, and state-level results formatting
+
 ## [0.0.1] - (Not yet published)
 
 - Initial release!